requirements.txt
absl-py==2.1.0
accelerate==0.34.2
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blis==0.7.11
bokeh==3.5.2
branca==0.7.2
cachetools==5.5.0
catalogue==2.0.10
catboost==1.2.7
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.0.0
cmdstanpy==1.2.4
confection==0.1.5
contourpy==1.3.0
convertdate==2.4.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dask==2024.8.2
dask-expr==1.1.13
distributed==2024.8.2
Django==5.1.1
djangorestframework==3.15.2
ephem==4.1.5
filelock==3.16.0
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.53.1
frozendict==2.4.4
fsspec==2024.9.0
gast==0.6.0
gensim==4.3.3
geopandas==1.0.1
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphviz==0.20.3
grpcio==1.63.2
grpcio-status==1.63.2
grpcio-tools==1.63.2
h5py==3.11.0
holidays==0.56
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.24.7
idna==3.8
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.7
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.5.0
lime==0.2.0.1
llvmlite==0.43.0
locket==1.0.0
LunarCalendar==0.0.9
lxml==5.3.0
lz4==4.3.3
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.1.0
multitasking==0.0.11
murmurhash==1.0.10
mysqlclient==2.2.4
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
param==2.1.1
partd==1.4.2
patsy==0.5.6
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.3
plotly==5.24.1
preshed==3.0.9
prophet==1.1.5
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyMeeus==0.5.12
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
pyspark==3.5.2
pystan==2.19.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
setuptools-git==1.2
shap==0.46.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
smart-open==7.0.4
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
stability-sdk==0.8.6
stanio==0.5.1
statsmodels==0.14.2
sympy==1.13.2
tblib==3.0.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.17.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.19.1
toolz==0.12.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
tweepy==4.14.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
xgboost==2.1.1
xyzservices==2024.9.0
yfinance==0.2.43
zict==3.0.0
zipp==3.20.2
